Separate the wallet in its own source file
